Add ScaleInfo header, remove redundant enums
