+++ SADXModLoader/SADXModLoader.vcxproj.filters
@@ -182,6 +182,9 @@
     <ClInclude Include="include\SADXEnumsNew.h">
       <Filter>Header Files\include</Filter>
     </ClInclude>
+    <ClInclude Include="include\ScaleInfo.h">
+      <Filter>Header Files\include</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="SADXModLoader.rc">
